--- src/core/tools/base/types.ts
@@ -1,7 +1,0 @@
-export type ToolInput<T> = T
-
-export interface ToolOutput<T> {
-    success: boolean
-    data?: T
-    error?: string
-}
