+++ src/core/tools/posthog/feature-flags/CreateFeatureFlagTool.ts
@@ -1,27 +1,13 @@
-import { z } from 'zod'
 import { PostHogTool } from '../PostHogTool'
 import type { ToolOutput } from '../../base/types'
 import type { ToolUse } from '../../../assistant-message'
-import { BasePostHogToolConfigSchema } from '../schema'
+import type { CreateFeatureFlagToolInput, CreateFeatureFlagToolOutput } from '../../schema'
+import {
+    CreateFeatureFlagToolInputSchema,
+    CreateFeatureFlagToolOutputSchema,
+    BasePostHogToolConfigSchema,
+} from '../../schema'
 
-export const CreateFeatureFlagToolInputSchema = z.object({
-    body: z.object({
-        key: z.string(),
-        name: z.string(),
-        active: z.boolean(),
-    }),
-})
-
-export type CreateFeatureFlagToolInput = z.infer<typeof CreateFeatureFlagToolInputSchema>
-
-export const CreateFeatureFlagToolOutputSchema = z.object({
-    id: z.number(),
-    name: z.string(),
-    key: z.string(),
-    active: z.boolean(),
-})
-
-export type CreateFeatureFlagToolOutput = z.infer<typeof CreateFeatureFlagToolOutputSchema>
 export class CreateFeatureFlagTool extends PostHogTool<CreateFeatureFlagToolInput, CreateFeatureFlagToolOutput> {
     autoApprove = false
     name = 'create_feature_flag'
