--- src/core/tools/posthog/schema.ts
@@ -1,7 +1,0 @@
-import { z } from 'zod'
-
-export const BasePostHogToolConfigSchema = z.object({
-    posthogApiKey: z.string(),
-    posthogHost: z.string(),
-    posthogProjectId: z.string(),
-})
